test(backend): reorganize
